requirements.txt
amqp==2.2.1
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
click==5.1
data-wizard==1.0.1.dev0
Django==1.11.4
djangorestframework==3.6.4
html-json-forms==1.0.0
idna==2.6
kombu==4.1.0
natural-keys==1.2.0
pkg-resources==0.0.0
pystache==0.5.4
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
redis==2.10.6
requests==2.18.4
six==1.10.0
urllib3==1.22
vine==1.1.4
wq.core==1.0.0
wq.io==1.0.0
xlrd==1.1.0
